insightcopilot/requirements.txt
fastapi
uvicorn
langchain
openai
pypdf2
python-dotenv
pandas
numpy
faiss-cpu
tiktoken